# website/api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"aisyndicate.infostrategic.com","basePath":"/","tags":[{"name":"investor-controller","description":"Investor Controller"},{"name":"startup-controller","description":"Startup Controller"}],"paths":{"/investor":{"post":{"tags":["investor-controller"],"summary":"Create investor","operationId":"createInvestorUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"investorRegistrationDto","description":"investorRegistrationDto","required":true,"schema":{"$ref":"#/definitions/InvestorRegistrationDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/InvestorDto"}},"201":{"description":"Created"},"400":{"description":"Investor with this email already exists"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["investor-controller"],"summary":"Update investor","operationId":"updateInvestorUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"investorDto","description":"investorDto","required":true,"schema":{"$ref":"#/definitions/InvestorDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/InvestorDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Investor with with specified ID doesn't exist"}}}},"/investor/assign":{"put":{"tags":["investor-controller"],"summary":"Assign startup to investor","operationId":"assignStartupUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"investorId","in":"query","description":"investorId","required":true,"type":"string","format":"uuid"},{"name":"startupId","in":"query","description":"startupId","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/InvestorDto"}},"201":{"description":"Created"},"400":{"description":"Startup is already assigned"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Investor or startup with with specified ID doesn't exist"}}}},"/investor/{id}":{"get":{"tags":["investor-controller"],"summary":"Get investor by id","operationId":"getInvestorUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/InvestorDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Investor with with specified ID doesn't exist"}}},"delete":{"tags":["investor-controller"],"summary":"Delete investor by id","operationId":"deleteInvestorUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/InvestorDto"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Investor with with specified ID doesn't exist"}}}},"/startups":{"get":{"tags":["startup-controller"],"summary":"Get all startups","operationId":"getAllStartupsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/StartupDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["startup-controller"],"summary":"Creates startup","operationId":"createStartupUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"startupDto","description":"startupDto","required":true,"schema":{"$ref":"#/definitions/StartupDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/StartupDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["startup-controller"],"summary":"Updates startup","operationId":"updateStartupUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"startupDto","description":"startupDto","required":true,"schema":{"$ref":"#/definitions/StartupDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/StartupDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Startup with specified ID doesn't exist"}}}},"/startups/create":{"post":{"tags":["startup-controller"],"summary":"Creates N random startupssss","operationId":"createNStartupsUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"amount","in":"query","description":"amount","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/StartupDto"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/startups/{id}":{"get":{"tags":["startup-controller"],"summary":"Get startup by id","operationId":"getStartupUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/StartupDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Startup with specified ID doesn't exist"}}},"delete":{"tags":["startup-controller"],"summary":"Deletes startup by id","operationId":"deleteStartupUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/StartupDto"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Startup with specified ID doesn't exist"}}}}},"definitions":{"InvestorDto":{"type":"object","properties":{"city":{"type":"string"},"country":{"type":"string"},"email":{"type":"string"},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"phoneNumber":{"type":"string"},"startups":{"type":"array","uniqueItems":true,"items":{"$ref":"#/definitions/StartupDto"}}},"title":"InvestorDto"},"InvestorRegistrationDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"title":"InvestorRegistrationDto"},"StartupDto":{"type":"object","properties":{"currentStatus":{"type":"string"},"description":{"type":"string"},"domain":{"type":"string"},"id":{"type":"string","format":"uuid"},"targetBudget":{"type":"integer","format":"int32"},"title":{"type":"string"}},"title":"StartupDto"}}}